+++ src/index.ts
@@ -1,3 +1,44 @@
 import { AvanzaClient, AvanzaClientOptions } from "./client/AvanzaClient";
+import type {
+  AccountType,
+  AccountsOverviewResponse,
+  AuthenticateBankIDResponse,
+  AuthenticateBankIDStatusResponse,
+  AuthenticateBankIDStatusResponseActive,
+  AuthenticateBankIDStatusResponseComplete,
+  AuthenticationSessionTOPTResponse,
+  BankIDAuthState,
+  InstrumentPosition,
+  InstrumentType,
+  Position,
+  PositionResponse,
+} from "./client/types";
+import type {
+  SearchResponse,
+  Hit,
+  TopHit,
+  Link,
+  PageSearchResults,
+} from "./client/market_types";
+
 export { AvanzaClient };
 export type { AvanzaClientOptions };
+export type {
+  AccountType,
+  AccountsOverviewResponse,
+  AuthenticateBankIDResponse,
+  AuthenticateBankIDStatusResponse,
+  AuthenticateBankIDStatusResponseActive,
+  AuthenticateBankIDStatusResponseComplete,
+  AuthenticationSessionTOPTResponse,
+  BankIDAuthState,
+  InstrumentPosition,
+  InstrumentType,
+  Position,
+  PositionResponse,
+  SearchResponse,
+  Hit,
+  TopHit,
+  Link,
+  PageSearchResults,
+};